+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 ## [Planned]
 - Speichervergleich BigDecimal zu PackedDecimal
 
+##[Unreleased]
+### Fixed
+- fehlerhaftes Caching von Zinssatz im bank-Package korrigiert
+- fehlerhaftes Caching von SNOMED im med-Package korrigiert
 
 ## [6.2.1] - 2025-05-27
 ### Fixed